LPS-71614 SF Rename
